third_party/blink/web_tests/platform/generic/editing/pasteboard/paste-list-003-expected.txt

Copy/pasting list items in a list. This test should be run with DRT for pasteboard access.

PASS

one: <b>one</b>
two: <span style="background-color: green">two</span>
three: three
one: <b>one</b>
two: <span style="background-color: green;">two</span>
four: <span style="background-color: orange"><i>four</i></span>
i love counting, counting to the number four: i love counting, counting to the number four
